refactor(gameMessage): 메시지 상수 constant에서 분리
